requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
berserker-resolver==2.0.1
boto3==1.17.5
botocore==1.20.5
branca==0.4.2
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
crypto==1.4.1
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
dnspython==1.15.0
dnspython3==1.15.0
docopt==0.6.2
filelock==3.0.12
folium==0.12.1
futures==3.1.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.1.0
html5lib==1.1
idna==2.10
importlib-metadata==2.1.1
ipwhois==1.2.0
ipython==7.20.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
maxminddb==2.0.3
mechanize==0.4.5
multidict==5.0.2
Naked==0.1.31
numpy==1.19.3
pandas==1.1.4
parsel==1.6.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.16
pycares==3.1.1
pycparser==2.20
pydig==0.3.0
PyDispatcher==2.0.5
pydnsbl==1.1.2
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.4.1
requests==2.7.0
requests-file==1.5.1
s3transfer==0.3.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
search-engines==0.5
shellescape==3.8.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
stem==1.8.0
Sublist3r==1.0
threadpoolctl==2.1.0
tld==0.12.3
tldextract==3.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
xortool==1.0.1
yarl==1.6.3
zipp==3.4.0
